refactor: adjust services into its own packages
