+++ src/main/java/com/jpcchaves/adotar/application/service/auth/contracts/SecurityContextService.java
@@ -1,4 +1,4 @@
-package com.jpcchaves.adotar.application.service.usecases;
+package com.jpcchaves.adotar.application.service.auth.contracts;
 
 import com.jpcchaves.adotar.domain.model.User;
