+++ src/main/java/com/jpcchaves/adotar/application/service/pet/contracts/PetCardService.java
@@ -1,4 +1,4 @@
-package com.jpcchaves.adotar.application.service.usecases;
+package com.jpcchaves.adotar.application.service.pet.contracts;
 
 public interface PetCardService {
   byte[] generatePetCard(Long petId);
