+++ src/main/java/com/jpcchaves/adotar/application/service/pet/contracts/PetCharacteristicService.java
@@ -1,6 +1,7 @@
-package com.jpcchaves.adotar.application.service.usecases;
+package com.jpcchaves.adotar.application.service.pet.contracts;
 
 import com.jpcchaves.adotar.application.dto.pet.PetCharacteristicsDto;
+
 import java.util.List;
 
 public interface PetCharacteristicService {
