+++ src/main/java/com/jpcchaves/adotar/application/service/user/UserServiceImpl.java
@@ -4,7 +4,7 @@ import com.jpcchaves.adotar.application.dto.ApiMessageResponseDto;
 import com.jpcchaves.adotar.application.dto.user.UserDetailsDto;
 import com.jpcchaves.adotar.application.dto.user.UserPictureDto;
 import com.jpcchaves.adotar.application.dto.user.UserUpdateNameDTO;
-import com.jpcchaves.adotar.application.service.usecases.SecurityContextService;
+import com.jpcchaves.adotar.application.service.auth.contracts.SecurityContextService;
 import com.jpcchaves.adotar.application.service.user.contracts.UserService;
 import com.jpcchaves.adotar.application.utils.user.UserUtils;
 import com.jpcchaves.adotar.domain.exception.UnexpectedErrorException;
