+++ src/main/java/com/jpcchaves/adotar/presentation/controller/v1/PetCharacteristicsController.java
@@ -1,14 +1,16 @@
 package com.jpcchaves.adotar.presentation.controller.v1;
 
 import com.jpcchaves.adotar.application.dto.pet.PetCharacteristicsDto;
-import com.jpcchaves.adotar.application.service.usecases.PetCharacteristicService;
+import com.jpcchaves.adotar.application.service.pet.contracts.PetCharacteristicService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
+
 import java.util.List;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -32,29 +34,29 @@ public class PetCharacteristicsController {
       summary = "Gets a list of pet characteristics",
       description = "Get a list of pet characteristics",
       responses = {
-        @ApiResponse(
-            description = "Success",
-            responseCode = "200",
-            content =
-                @Content(
-                    array =
-                        @ArraySchema(
-                            schema =
-                                @Schema(
-                                    implementation =
-                                        PetCharacteristicsDto.class)))),
-        @ApiResponse(
-            description = "Bad Request",
-            responseCode = "400",
-            content = @Content),
-        @ApiResponse(
-            description = "Unauthorized",
-            responseCode = "401",
-            content = @Content),
-        @ApiResponse(
-            description = "Internal Error",
-            responseCode = "500",
-            content = @Content),
+          @ApiResponse(
+              description = "Success",
+              responseCode = "200",
+              content =
+              @Content(
+                  array =
+                  @ArraySchema(
+                      schema =
+                      @Schema(
+                          implementation =
+                              PetCharacteristicsDto.class)))),
+          @ApiResponse(
+              description = "Bad Request",
+              responseCode = "400",
+              content = @Content),
+          @ApiResponse(
+              description = "Unauthorized",
+              responseCode = "401",
+              content = @Content),
+          @ApiResponse(
+              description = "Internal Error",
+              responseCode = "500",
+              content = @Content),
       })
   public ResponseEntity<List<PetCharacteristicsDto>> getAll() {
     return ResponseEntity.ok(petCharacteristicService.getAll());
